new changes

+++ gestionControleDAcces/src/main/java/oprag/project/gestionControleDAcces/services/InspectionService.java
@@ -8,7 +8,7 @@ public interface InspectionService {
     InspectionDAO save(InspectionDAO inspectionDAO);
     List<InspectionDAO> findAll();
     InspectionDAO findById(Integer id);
-    List<Long> pieChartStatsData();
+    List<Double> pieChartStatsData();
     InspectionDAO findByNom(String nom);
     void delete(Integer id);
 }
